+++ src/op/degree_sequence.rs
@@ -20,11 +20,6 @@
 //! assert!(digraph.degree_sequence().eq([3, 2, 3]));
 //! ```
 
-use crate::{
-    Degree,
-    Vertices,
-};
-
 /// Digraph degree sequence
 pub trait DegreeSequence {
     /// Iterate the digraph's degree sequence.
@@ -50,13 +45,4 @@ pub trait DegreeSequence {
     /// ```
     #[must_use]
     fn degree_sequence(&self) -> impl Iterator<Item = usize>;
-}
-
-impl<D> DegreeSequence for D
-where
-    D: Degree + Vertices,
-{
-    fn degree_sequence(&self) -> impl Iterator<Item = usize> {
-        self.vertices().map(move |v| self.degree(v))
-    }
 }