+++ src/repr/adjacency_map/mod.rs
@@ -115,6 +115,8 @@ use {
         Complete,
         Converse,
         Cycle,
+        Degree,
+        DegreeSequence,
         EdgeList,
         Empty,
         ErdosRenyi,
@@ -431,6 +433,12 @@ impl Cycle for AdjacencyMap {
     }
 }
 
+impl DegreeSequence for AdjacencyMap {
+    fn degree_sequence(&self) -> impl Iterator<Item = usize> {
+        self.vertices().map(move |v| self.degree(v))
+    }
+}
+
 impl Empty for AdjacencyMap {
     /// # Panics
     ///
@@ -587,7 +595,7 @@ impl GrowingNetwork for AdjacencyMap {
 
 impl HasArc for AdjacencyMap {
     fn has_arc(&self, u: usize, v: usize) -> bool {
-        self.arcs.get(&u).map_or(false, |set| set.contains(&v))
+        self.arcs.get(&u).is_some_and(|set| set.contains(&v))
     }
 }
 
@@ -762,7 +770,7 @@ impl RandomTournament for AdjacencyMap {
 
 impl RemoveArc for AdjacencyMap {
     fn remove_arc(&mut self, u: usize, v: usize) -> bool {
-        self.arcs.get_mut(&u).map_or(false, |set| set.remove(&v))
+        self.arcs.get_mut(&u).is_some_and(|set| set.remove(&v))
     }
 }